refactor: move log_memory_usage to new log module in utils
