+++ crates/proof-of-sql/src/utils/log.rs
@@ -1,0 +1,30 @@
+use sysinfo::System;
+use tracing::{trace, Level};
+
+/// Logs the memory usage of the system at the TRACE level.
+///
+/// This function logs the available memory, used memory, and the percentage of memory used.
+/// It only logs this information if the TRACE level is enabled in the tracing configuration.
+///
+/// # Arguments
+///
+/// * `name` - A string slice that holds the name to be included in the log message.
+#[allow(clippy::cast_precision_loss)]
+pub fn log_memory_usage(name: &str) {
+    if tracing::level_enabled!(Level::TRACE) {
+        let mut system = System::new_all();
+        system.refresh_memory();
+
+        let available_memory = system.available_memory() as f64 / (1024.0 * 1024.0);
+        let used_memory = system.used_memory() as f64 / (1024.0 * 1024.0);
+        let percentage_memory_used = (used_memory / (used_memory + available_memory)) * 100.0;
+
+        trace!(
+            "{} Available memory: {:.2} MB, Used memory: {:.2} MB, Percentage memory used: {:.2}%",
+            name,
+            available_memory,
+            used_memory,
+            percentage_memory_used
+        );
+    }
+}
