+++ crates/proof-of-sql/src/utils/mod.rs
@@ -1,3 +1,6 @@
 //! This module contains utilities for working with the library
 /// Parse DDLs and find bigdecimal columns
 pub mod parse;
+
+/// This module provides logging utilities for the library, including functions to log system memory usage.
+pub mod log;
